requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
earthaccess==0.12.0
xarray==2023.11.0
netcdf4==1.6.5
h5netcdf==1.3.0
numpy==1.24.4
scipy==1.11.4
pandas==2.1.3
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
rasterio==1.3.9
folium==0.15.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
